Add named routes and 'Router.url_for' utility
For 'Route' implementation that support reversing URLs, an extra 'name'
parameter has been added to their respective helpers in 'Router'. The
route is indexed and can be later retreived by 'Router.url_for' to
generate an URL.

The utility only produce a corresponding pattern, but a full URL, which
could be performed with 'Soup.URI'.