vsgi: Improve logging and support structured logging
Add a '--log-writer' argument to specify a writer for the logs.
Supported values are:

 - standard-streams
 - journald
 - default

If glib-2.0 (>=2.50) is available, then use the corresponding log writer
functions.

Add an optional dependency on 'libsystemd' as a fallback for
'Log.writer_journald' if glib-2.0 (>=2.50) is not satisfied.

Move the colored output as a fallback for 'standard-streams' if
glib-2.0 (>=2.50) is not satisfied.