vsgi: Mitigate blocking by dispatching asynchronously with 'IOSchedulerJob'
Dispatch incoming requests using 'IOSchedulerJob', which process using
a small internal pool of threads.

This is only a temporary solution before we move on a fully asynchronous
model. It is assuming that processing requests is a relatively short
operation.

There's no need to use 'Idle.add' in the CGI implementation since the
initial request will be triggered once the main loop starts.